+++ mobile/shared/components/geckoview/GeckoViewStartup.sys.mjs
@@ -219,6 +219,14 @@ export class GeckoViewStartup {
           ],
         });
 
+        GeckoViewUtils.addLazyGetter(this, "GeckoViewPreferences", {
+          module: "resource://gre/modules/GeckoViewPreferences.sys.mjs",
+          ged: [
+            "GeckoView:Preferences:RegisterObserver",
+            "GeckoView:Preferences:UnregisterObserver",
+          ],
+        });
+
         break;
       }